RT #43314 Modules::RequireExplicitInclusion with __PACKAGE__ RT #44609 Support for __PACKAGE__
RT #44609 included a patch, but this was not used because it would also
have caused the policy to accept __PACKAGE__::foo() and
$__PACKAGE__::foo, neither of which is correct. Adding '__PACKAGE__' to
@builtin_packages has the same problem.

Since adding hooks for type-specific filtering seemed not to be
warranted, I simply had _find_class_method_calls ignore calls where the
invocant is '__PACKAGE__'. A lot of comment verbage was splashed into
the code at that point to call attention to this little hack.

The t/Modules/RequireExplicitInclusion.run file got two more tests
added; one to be sure __PACKAGE__->foo() is allowed, and one to make
sure things like __PACKAGE__::foo() and $__PACKAGE__::foo are not.

In theory someone could write a __PACKAGE__ package, and in theory if
this is included __PACKAGE__::foo() and friends will work, but I did not
check for it.

Yes, the Changes file was updated.