RT #43310: Subroutines::ProhibitCallsToUndeclaredSubs with version numbers in import spec
The problem was that the policy was unable to find the imported
subroutines in

use HTTP::Status '5.817' => qw(HTTP_INTERNAL_SERVER_ERROR);

because it made the assumption that if the first token after the
imported module name was a PPI::Token::Quote, then all the other tokens
of interest were also.

This commit changes this assumption to a scan of the arguments looking
for lists, quotes, and qw{} strings, assuming all these are of interest
and nothing else is. This will pick up the version in the example above,
but so did the original code.