RT #72022 - ProhibitCallsToUndeclaredSubs reports loop labels as violations
This was handled in
Perl::Critic::StricterSubs::Utils::find_subroutine_calls() in a manner
strictly analogous to (i.e. cribbed shamelessly from) the handling of
bareword filehandles. Some tests, taken from the RT ticket, were added
to the .run file.

The patch handles next, last, and redo, which only take labels. It does
not handle goto, which takes either a label or an expression.